requirements_streamlit.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
openai>=1.0.0
